requirements.txt
apify>=2.0.0
crawlee>=0.3.0
requests>=2.31.0
readability-lxml>=0.8.1
markdownify>=0.11.6
lxml>=4.9.0
